app/requirements.txt
cryptography==39.0.2
aiohttp==3.8.1
pytest==7.1.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-env==0.6.2
alembic==1.10.2
sqlalchemy==2.0.6
marshmallow==3.19.0
aiohttp_cors==0.7.0
requests==2.28.2
pyparsing==3.0.9
pyyaml==6.0
asyncpg==0.27.0
aiohttp_session==2.12.0
aiohttp_apispec==2.2.3
aioresponses==0.7.4
iniconfig==2.0.0
black==23.1.0
aio-pika==9.0.5
fabric==3.0.0
